backend/requirements.txt
# === Core API stack ===
fastapi==0.115.2
uvicorn==0.31.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
python-multipart==0.0.9
boto3==1.35.39
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
google-api-python-client==2.153.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
pydantic-core==2.23.4
# === LangChain + Anthropic (stable & compatible) ===
langchain==0.2.16
langchain-core==0.2.40
langchain-community==0.2.11
langchain-groq==0.1.6

openai>=1.3.0
